requirements.txt
selenium
chromedriver-autoinstaller
pyqt5
pandas
request
